UHF-9832: Added a separate test for the user sanitize drush command.
